04-04-16-requirements.txt
Django==1.6.11
Django-Select2==4.3.2
-e git+https://github.com/Geonode/geonode.git@8dc5e2f341c5804c26b811515c6bb2fb0364a1ee#egg=GeoNode-2.4
Markdown==2.6.6
MultipartPostHandler==0.1.0
OWSLib==0.10.0
Paver==1.2.1
Pillow==3.1.1
Shapely==1.3.1
South==1.0.2
URLObject==2.4.0
Unidecode==0.04.19
YURL==0.13
aldryn-apphooks-config==0.2.7
aldryn-boilerplates==0.7.4
aldryn-common==1.0.1
aldryn-search==0.2.11
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
awesome-slugify==1.6.2
beautifulsoup4==4.2.1
billiard==3.3.0.23
celery==3.1.17
cmsplugin-filer==0.10.1
django-activity-stream==0.4.5beta1
django-admin-enhancer==0.1.3.1
django-appconf==0.6
django-appdata==0.1.5
django-autocomplete-light==2.2.10
django-bootstrap3-datetimepicker==2.2.3
django-celery==3.1.16
django-classy-tags==0.7.2
django-cms==3.1.5
django-debug-toolbar==1.0
django-downloadview==1.2
django-extensions==1.2.5
django-filer==0.9.8
django-forms-bootstrap==3.0.1
django-formtools==1.0
django-friendly-tag-loader==1.1
django-geoexplorer==4.0.5
django-guardian==1.2.0
django-haystack==2.1.0
django-jsonfield==0.9.12
django-leaflet==0.13.7
django-meta==1.0
django-meta-mixin==0.3.0
django-modeltranslation==0.8
django-mptt==0.6.1
django-nose==1.2
django-nyt==0.9.9
django-pagination==1.0.7
django-parler==1.6.2
django-polymorphic==0.5.6
django-reversion==1.8.0
django-sekizai==0.8.2
django-smart-load-tag==0.3.2
django-sortedm2m==1.1.1
django-spurl==0.6.4
django-standard-form==1.1.1
django-taggit==0.18.1
django-taggit-autosuggest==0.2.4
django-taggit-templatetags==0.2.5
django-tastypie==0.11.0
django-templatetag-sugar==1.0
django-treebeard==3.0
djangocms-admin-style==1.1.0
djangocms-apphook-setup==0.1.2
djangocms-blog==0.5.0
djangocms-link==1.7.2
djangocms-text-ckeditor==2.3.0
easy-thumbnails==1.4
flake8==2.3.0
geolinks==0.0.1
geonode-agon-ratings==0.3.1
geonode-announcements==1.0.5
geonode-arcrest==10.2
geonode-avatar==2.1.4
geonode-dialogos==0.4
geonode-notification==1.1.1
geonode-user-accounts==1.0.10
geonode-user-messages==0.1.2
gisdata==0.5.4
gsconfig==1.0.3
gsimporter==1.0.0
html5lib==0.999
httplib2==0.8
ipython==2.2.0
kombu==3.0.35
lxml==3.6.0
mccabe==0.4.0
nose==1.3.1
numpy==1.10.4
pep8==1.6.2
pinax-theme-bootstrap==3.0a11
pinax-theme-bootstrap-account==1.0b2
psycopg2==2.6.1
pycsw==1.10.3
pyelasticsearch==0.6.1
pyflakes==1.1.0
pygdal==1.10.0.0
pyproj==1.9.5.1
python-dateutil==2.5.1
python-mimeparse==1.5.1
pytz==2016.3
regex==2016.03.02
requests==2.9.1
simplejson==3.8.2
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.1.19
transifex-client==0.10
wiki==0.1
wsgiref==0.1.2
